--- x-pack/plugins/lens/public/visualizations/reveal_image/dimension_editor.scss
@@ -1,3 +1,0 @@
-.lnsDynamicColoringRow {
-  align-items: center;
-}
